LPS-79878 This verification was added by LPS-27765 in 6.1. It is only needed for upgrades to 6.2
